Theme Provider is separated from config provider
