+++ src/components/use-config/index.ts
@@ -1,4 +1,4 @@
 export { useConfigs, defaultToastLayout } from './config-context';
-export { default as ThemeProvider } from './theme-provider';
+export { default as ThemeProvider } from '../theme/theme-provider';
 
 export type { UpdateToastsFunction, UpdateToastsLayoutFunction, UpdateToastsIDFunction } from './config-context';
